requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
streamlit==1.39.0
pymupdf==1.24.10
spacy==3.7.6
sentence-transformers==3.1.1
faiss-cpu==1.8.0
openai==1.47.0
python-dotenv==1.0.1
plotly==5.24.1
